HostBase handles commands properly.
MQTT is a singleton, whereas it was a client per HostBase in JavaScript.

MyQ Microservice is working for state coming from hubs and sensors.

Commands to open/close doors not implemented in the server.dart file yet.

The code in the MyQ package to command the doors open/closed has not been tested yet.